commands: move interdiff code to interdiff.rs
